Move InspectCodeImporter to own module
